requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
anyio==4.3.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.0
async-lru==2.0.5
attrs==23.2.0
babel==2.17.0
beautifulsoup4==4.14.2
bleach==6.3.0
blinker==1.8.2
blis==1.3.2
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.2.2
cffi==2.0.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.23.0
colorama==0.4.6
comm==0.2.3
confection==0.1.5
cymem==2.0.13
dataclasses-json==0.6.5
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
executing==2.2.1
faiss-cpu==1.12.0
fastjsonschema==2.21.2
fqdn==1.5.1
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
ipykernel==7.1.0
ipython==9.7.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.4
json5==0.12.1
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.10
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
langchain==0.1.19
langchain-community==0.0.38
langchain-core==0.1.52
langchain-experimental==0.0.58
langchain-openai==0.1.6
langchain-text-splitters==0.0.1
langchainhub==0.1.15
langsmith==0.1.56
lxml==6.0.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
matplotlib-inline==0.2.1
mdurl==0.1.2
mistune==3.1.4
multidict==6.0.5
murmurhash==1.0.15
mypy-extensions==1.0.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.4.7
notebook_shim==0.2.4
numpy==1.26.4
openai==1.27.0
orjson==3.10.3
overrides==7.7.0
packaging==23.2
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.5
pillow==10.3.0
platformdirs==4.5.0
preshed==3.0.11
prometheus_client==0.23.1
prompt_toolkit==3.0.52
protobuf==4.25.3
psutil==7.1.3
pure_eval==0.2.3
pyarrow==16.0.0
pycparser==2.23
pydantic==2.7.1
pydantic_core==2.18.2
pydeck==0.9.0
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-json-logger==4.0.0
pytz==2024.1
pywinpty==3.0.2
PyYAML==6.0.1
pyzmq==27.1.0
referencing==0.35.1
regex==2024.4.28
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.1
Send2Trash==1.8.3
six==1.16.0
smart_open==7.5.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.8
spacy==3.8.9
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.30
srsly==2.5.1
stack-data==0.6.3
streamlit==1.34.0
tenacity==8.3.0
terminado==0.18.1
thinc==8.3.9
tiktoken==0.6.0
tinycss2==1.4.0
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
typer-slim==0.20.0
types-requests==2.31.0.20240406
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
uri-template==1.3.0
urllib3==2.2.1
wasabi==1.1.3
watchdog==4.0.0
wcwidth==0.2.14
weasel==0.4.3
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
widgetsnbextension==4.0.15
wrapt==2.0.1
yarl==1.9.4
